[IMP] `restore`: pipe extract
The temp partitions size can be limited. The original solution to extract zipped dumps was using pythons `tempfile` lib which by standard writes to the temp partition. Especially big databases are able to fill the temp partition on many systems.

Why use the filesystem in the first place?
`dump.sql` is directly piped to `psql`. The filestore is handled according to the users choice, also w/o `tempfile`.