[FIX] utils.github: resolve worktrees paths to check identity
The current logic in `git_worktree_create()` for ensuring a worktree path is checked out works by: attempting to add the worktree, and if that fails, assuming that if it appears in `git worktree list` it means it exists already. When comparing paths, we need to make sure all symlinks are resolved first, both in the worktree path we're trying to create and the paths returned by git.