[IMP] remove: allow removing non Odoo databases
When creating a database with the create command, the user might want
to remove it without initializing it. Currently that fails, raising an
Exception when getting the version with
`version = self.db_version_clean()`.

This commit fixes that and adds a warning message to the prompt when
deleting a non Odoo database.