[REF] Refactor self-update
Refactor the self-update mechanism to improve reliability
and performance while keeping the code readable.

The new mechanism is based on the following principles:

1.  When odev is run, check its local repository for changes against
    the remote tracked branch (not targeting a specific branch anymore),
    but only if the check interval has passed since the last check.
2.  Add the possibility for the user to choose whether to download
    updates automatically, never download updates or ask before
    downloading.
3.  If an update is available and changes are pulled, check whether
    the requirements.txt file inside of the repo was modified. If so,
    install the new dependencies.

In order to save execution time, changes are not fetched in the current
process anymore but in a separate - detached - subprocess. This implies
that updates can be delayed until the next time odev is run but it has
the benefit of not slowing down the current execution.