[FIX] python.py: ignore special case for util package in requirements
`util_package` is added to odoo.sh projects as a submodule with the
help of `requirements.txt`, where a line is added that points to a
filepath specific for odoo.sh containers. This makes the `pip install`
command return a non-zero exit code because of the none existant path
on the local machine.

Add a special case to check for this until this process changes.