[IMP] Add handling of command runtime errors
Add a new CommandError class that is caught when a command is run. This
allows to handle errors in a more generic way, and to display
an error log instead of a stack trace.

Add a new method `error()` to the BaseCommand class to generate
a CommandError without having the need to import the class.