[IMP] connectors.postgres: add helper methods and improve caching
LRU caching from functools was used to cache all SQL queries. This
was not a good idea because it was not possible to invalidate the
cache and it was not efficient because of context managers which made
the `self` object to be generated each time a new connection was open.

The commit adds a simple cache system which can be bypassed easily
by setting the `nocache` attribute to `True`. Only `SELECT` statements
are cached to let alteration statements to be executed at all times.