[ADD] Github connector and secrets vault
New connector to the Github API and to manage local Git repositories.
Handles the clone and pull of repositories, and the creation of
worktrees.

The secrets vault is a new module to manage credentials and secrets
in a secure way. It uses the the local SSH agent to encrypt secrets
and stores them in a local PostgreSQL database.