[IMP] Synchronize repositories and worktrees if deleted from elsewhere
When a repository is removed from the filesystem `odev` will attempt
to clone it anew. This might lead to errors when creating worktrees
if the old worktree directory is still present. This commit checks
for existing worktrees and removes them before cloning the repository.

When a worktree is removed from the filesystem but is still referenced
in Git, `odev` will attempt to load and use it. This might lead to
errors where we are loading a non-existing path. This commit checks
for existing worktrees references and removes them before re-creating
the worktree.

Because fetching changes in repositories is done in a detached process,
`odev` does not wait for the process to end before running the next
operation. If a call to :func:`fetch` or :func:`fetch_worktrees` is
followed by a call to :func:`pull` or :func:`pull_worktrees`, the fetch
subprocess might still be running when the pull is executed. This leads
to situations where the pull is executed on a repository that is not
entirely fetched yet and Git raises an error stating that rebase cannot
be performed on two different branches at once. This commit inverts
calls to fetch and pull so that the fetch is always executed *after* the
call to git pull. Changes pulling might be delayed until the next time
`odev` is run.