[IMP] SQL queries context
Run SQL queries in transactions and rollback if necessary

Make sure SQL queries are surrounded by `BEGIN` and `COMMIT` statements
(or `ROLLBACK` if an error ocurred).

Queries are now run in a new thread to allow cancelling long running
operations (i.e. creating a database from a large template). This is
an improvement for security and user experience, not performance hence
the thread is called in a blocking way and no code is executed
in parallel in the parent thread.