[IMP] Framework improvements:
Split `PostgresDatabase` into `LocalDatabase` and `PostgresDatabase`
to allow reusability of the `PostgresDatabase` class for other
non-Odoo databases.

Make the Odev framework global and accessible from anywhere through
the `OdevFrameworkMixin` mixin class. This adds easy to access
properties for the configuration and the datastore.

Move datastore models to a new store instance with a more consistent way
of creating and accessing tables.

Harmonize the use of properties instead of methods when possible.