[ADD] commands.database.test: Run tests on an empty copy of the database
Copy the selected database configuration and initialize a new database
with no data. Install the selected modules and run unit tests.
Capture the output and display it in the terminal and keep track
of failing tests to display a summary at the end of the command.
Delete the test database after the tests have been run.

Odoo-bin and other python scripts can be run with a progress callback
that is called on every line outputted to STDOUT or STDERR
by the script.