[IMP] common.progress: prevent clearing lines when live display is on
When a live display is running and clearing lines, ANSI escape codes
are appended to the stream buffer and not flushed properly. This leads
to a mess of escape codes in the terminal after the end of the live
display because ANSI codes are rendered at once without their escape
sequence.

When in debug mode (`--log-level=debug`), the live display is disabled
altogether.