[REF] common.console: refactor prompt and style into console module
Replace existing modules `odev.common.style` and `odev.common.prompt`
with a new module `odev.common.console` that provides a single
entry point for all interactions with the user through the terminal.

The new module provides a `Console` class that is instantiated and
shared across other modules so that they can share the same methods.

The refactor also aims to fix the live displays and line clearing
that used to clash and print improperly escaped ANSI codes in the
terminal.