[REF] connectors.rest: harmonize downloads progress
Let the base `RestConnector` class handle the progress bar for
downloaded files. This is done by adding a new `download` method
that takes care of the progress bar and the file writing.

Adds a download column to the progress to display the size downloaded
and the total size of the file.