[FIX #59] Bump minimal python version to 3.10 & config fixes
Update documentation to require python 3.10 or higher when installing
and running odev. This does not impact python versions used in the
virtual environments of managed odoo databases.

Replace instances of `StrEnum` with basic abstract classes as it was
introduced in python 3.11.

--- Fixes #59

Fix an issue where the config object was called as a context manager
without the `__enter__` and `__exit__` methods implemented.
As the config object is initialized during the load of the framework,
it is no longer necessary to use it in a context.