[IMP] plugins: add limited support for external commands
First step for adding support for external addons that can extend odev
features. This is a very limited support that only load commands from
external addons, with a command to load/unload plugins.

Plugins must be located in an external repository and must follow
a strict convention for the directory and file structure.