[FIX] list: colored circles on all terminals with fonts
Some terminals and font don't handle coloring emojis while others do but
don't handle the pre-colored version of the emoji itself. This commit
forces both the color and the colored version of the emoji so that it
works on all terminals with (mostly) all fonts.

Tested with SauceCode Pro Nerd Font.