[FIX] secrets: error during encryption of empty passwords
When encrypting a password, the encryption process was failing when the
password was empty. This was due to the fact that the encryption process
needs one or more characters to work on and empty strings are not
supported.

Empty passwords now return an empty string instead of a password, this
should not be an issue security-wise because empty passwords are
temporary and only stored as-is when the user is creating a new password
and entered their login but not their password yet.