[IMP] Improve plugins support with topological dependencies
Add support for a `__manifest__.py` file in the root of plugins,
allowing for transferring information from a plugin to odev
(description, version number and dependencies).

This allows for topological dependencies between plugins, and automatic
download of dependant plugins when installing a plugin.

The manifest file is a python file with the following structure:

```py
"""This docstring is used as the plugin description.

It can span over multiple lines and include [bold]format options[/bold]
for printing through the odev console.
"""

__version__ = "1.0.0"  # The version of the plugin
depends = ["plugin1"]  # The list of plugins this plugin depends on
```

Additionally, the `odev plugin` command has been improved to allow for
installing plugins with dependencies, and to display the description of
a plugin as well as its current status.