[IMP] tests: refactor `OdevTestCase` and improve patching
Improve `OdevTestCase` for easier patching of methods and attributes,
including classes not directly available under the `odev` instance.

Allow calling `Odev:dispatch` with forced arguments instead of reading
from the CLI at all times, allows for easier testing of commands
and a more thorough testing of the framework itself with results closer
to real-life situations.

Add a new class `OdevCommandTestCase` that extends `OdevTestCase` and
is used specifically for running commands.

Automatically enable capturing output when commands are run in tests,
including logs, stdout and stderr. This allows for easily checking
the whole output of a command when run.

Adapt existing tests to use the new classes and features.