[4.0][FIX] upgrade 4.0.20: ignore invalid agentcrypt header
Ignore passwords with an invalid agentcrypt header when upgrading to
4.0.20. The error likely happens because:

- the password is empty
- the password is not encrypted using agentcrypt

The upgrade script will log a debug message and continue with
the upgrade as if nothing happened. If a password is not properly
encrypted after the upgrade, the credentials will fail to authenticate
on next used and will be asked anew to the user.