[4.0][FIX] create: use correct venv on database creation
On the first initialization of a database for which the virtualenv does
not exist and with a specific Odoo version, the virtualenv is created
under `master` instead of the version number.

This commit fixes this issue by making sure running the `odoo-bin`
process with a version also overrides the virtualenv name and path.

Improve virtual environment saving and loading by correctly storing the
name of the virtual environment in the databases table. Also load
the stored virtual environment name when running a database.

Improve handling on the database class altogether, the `venv`
attribute now returning a `PythonEnv` instance instead of a path
to a virtual environment.

Improve tests and fix a bug in the base `OdevTestCase` class where
it was sometimes impossible to cleanup the database store.