[IMP] progress: inform when progress is disabled
When a debugger call is found in odev's source code, live status
(progression bars and spinners) are disabled to avoid any conflict
with I/O in the shell.

The locations of the debuggers and an informative message were logged
to the console in debug mode, which wasn't clear to the user. Those
debug logs are converted to warnings with a clear message to inform
the user that the progress is disabled. Spinners are replaced by an info
log message to keep informing the user of the progress.