[FIX] delete: no match deletes all databases
Running `odev delete <db>` was targeting all existing databases at once
if the database did not exist. This was due to the fact that the
command was not stopping after checking if the database existed, and
then proceeded with scanning all databases with no regular expression
and no other filter.

This is fixed by adding the database name to the predicate that filters
the databases to delete and by making the arguments `database` and
`expression` mutually exclusive.

Added unit tests for the delete command.