[4.0][IMP] odoobin: display print when streaming logs
When running `odoo-bin` through odev, the logs are streamed from
the subprocess and piped back into the CLI with postprocessing and
re-coloring of the logs in real-time. This has the drawback of missing
out on `print` statements that are not part of the logs.

This commit fixes this by replacing to call to `Popen` with a custom
subprocess that captures both stdout and stderr and pipes them back
in a pseudo-TTY.

This commit also improves the detection of calls to interactive
debuggers by externalizing the detection logic into a separate module.
This avoid repeating code multiple times and maintaining multiple regex
patterns sharing the same objective.