[IMP] git: handle worktrees and custom revisions
Manage worktrees within odev through a new command `odev worktree`
which allows for creating, removing, pruning and listing existing
worktrees, as well as changing the revisions of all repositories
independently from one-another.

A worktree can be specified with commands running odoo-bin and
the worktree will be save with the database for subsequent uses.
This allows overriding the revision of a repository for a specific
database without affecting the other databases nor corrupting
existing worktrees.

Additionally bring a few small improvements with the help
of time-sensitive cache for performance-hungry operations such as
listing installed pip dependencies.