[4.3][FIX] odev: only set the password of 50 users to 'odoo'
Previously, we would set the password of all users except the admin to
'odoo'. When updating a module on the database later, in some cases, all
passwords that are not yet hashed with pbkdf2 will be rehashed. This is
painfully slow for databases with many users. This commit fixes that by
only updating the passwords of 50 users.

An alternative solution would be to hardcode a pbkdf2 hash of 'odoo',
but this would look like a magic string. Might be considered in the
future if there is real utility in logging in as any user.