[IMP] run/test: cleanup streaming of odoo-bin logs
Simplify the code responsible of streaming and live-processing the logs
of sub-processes run with `bash.stream()`. Read characters one by one
until a newline character is found rather than reading an arbitrary
amount of bytes and splitting lines afterwards.

This change should not affect the behavior of the `bash.stream()` method
but should make it more robust and easier to understand.

Additionally, improve the the `odoobin_progress` methods for existing
commands to reuse part of the code in semi-private methods instead of
rewriting it multiple times across the codebase.