[IMP] workflows: test upgrade scripts
Add a GitHub workflow to test the upgrade scripts and process, making
sure odev updates and still works without issues after new releases.

Also fix an issue in the preparation of the datastore after changing
the structure of the `secrets` table. Changing the primary key of
the table is now supported in the preparation of the datastore ran at
every upgrade.

Miscellaneous other small improvements and fixes are included.