[IMP] odev.commands.{run,create}: run and create databases from template
Externalize template database detection in a new class
`OdoobinTemplateCommand` that is used by both `run` and `create`.

Rename the `--template` option to `--from-template` in both commands.

Allow the `--from-template` option to be used with a database name
or with no value to fallback to an existing template of the current
database.

---