[IMP] odoobin: database repository override
When running an odoo database if the current working directory is a git
repository it is used as the database repository. This is used to avoid
requiring the user to input the repository as an addon-path when running
the command. The repository is saved at first run to avoid asking the
user again on subsequent runs even if the command is called from another
working directory.

In that case, if the new working directory is also a git repository, the
repository linked to the database was silently changed to the new one.
This could lead to unexpected behavior and confusion, as running from
another repository could simply be an input error and changing
the addons paths of a database could lead to it not loading properly.

We address this issue by detecting if the current repository
is different from the one linked to the database and asking the user
to confirm the change, or whether the old repository should be used
instead.

If the user confirms the change, the repository is updated and saved
to the database. If the user chooses to keep the old repository, the
command is run with that previous addons-path and the linked repository
is kept as-is.