[4.5] Duplicate process in `bash.stream()` cannot be killed (#151)
* [4.5][IMP] workflows: fail tests faster

When a test fails, the workflow should fail as soon as possible to
avoid wasting time on the test workers.

* [4.5][IMP] tests: allow running odoo-bin in unit tests

Refactor part of the tests framework to allow running odoo-bin during
unit tests, by letting the tests download odoo's repositories
in a lightweight way.

Add safeguards to prevent interfering with the user's local files.

* [4.5][FIX] common.bash: duplicated process when streaming command

Resolve an issue where the process was duplicated when running a command
through `bash.stream()` because using `Popen` with the argument
`shell=True`, which is considered unsafe as it could lead to shell
injection attacks. This also had the side effect of forking a new shell
process and running the command in it, which was then preventing
interactions to terminate the process:

- Manually sending CTRL+C to the process was needed to be done twice
  to signal both processes to terminate
- Calling `odev kill <database>` was not terminating the process
  because the process being killed was the forked process and not
  the shell process, which was still running in blocking mode

With this commit, the `Popen` call is done with `shell=False` and the
command is passed as a list of arguments, which is the recommended way
as per the Python documentation.

Upon reception of a CTRL+C char from the user, the process is now
terminated using `Popen.terminate()` rather than relaying the control
character to the subprocess.

* [4.5][FIX] commands.create/test: fixes on version and venv detection

Correct the version and virtual environment detection for databases
and commands. Allow calling `odev create` without specifying the
version, and use the default version "master".

* [4.5][FIX] connectors.git: add support for remote revisions

Allow creating worktrees from remote branches.

* [4.5][IMP] common.bash: mimic streaming on non-TTY shells

Execute the command instead of streaming it when the shell is not a TTY.