[FIX] common.commands.odoobin: Escape markup in odoo-bin logs
Escape Rich's markup-like substrings in odoo-bin logs to avoid having
information stripped from the logs when displayed in the console.

For example, the following log line:
```
Element '<xpath expr="//field[@name=&#39;test&#39;]">' cannot be located in parent view
```

Was previously stripped down to:
```
Element '<xpath expr="//field">' cannot be located in parent view
```

Additionally, a new argument is passed to the odoo-bin command
and children to allow disabling pretty-printing of the logs.