[FIX] odev: parsing of unknown arguments in python 3.12.7+
Breaking change: passing unknown (odoo-bin) arguments before positional
arguments is not supported in Python >= 3.12.7. In the same idea, it is
also not supported anymore to pass known arguments after positional
arguments.

It is not planned to reimplement this feature in the later versions
as it was taking advantage of a bug not compliant with common CLI
rulings and RFCs.

When in doubt, please refer to the command's help and follow
the proposed order for arguments.

See: https://github.com/python/cpython/issues/59317