[FIX] setup: symlink creation falls back under `/usr/local/bin`
Odev used to creat symlinks under `~/.local/bin` which is not in
the PATH by default on some operating systems. While it has
the advantage of not requiring root privileges to install the odev CLI
tool, the command remains unavailable to the user until they add
`~/.local/bin` to their PATH manually. This was observed on Ubuntu.

This commit changes the symlink creation to `/usr/local/bin` if
the first path is not currently available in the path.
`/usr/local/bin` is in the PATH of all Unix-based systems by default.
When privileged accesses are required, the user is prompted to enter
their password which is then passed through transparently and securely
to the operating system.