[FIX] odev: Neutralization fails after dump restoration
When restoring a backup, the neutralization process fails because of
a missing process attribute set on the new local database instance.

The missing process seems to be due to a race condition between the
python process of Odev and the underlying PostgreSQL database server.
The restored database is still pending transactions or commits for
a while when the neutralization process starts.

This commit adds a new loop around the restore subcommand to wait for
the database to be ready before starting the neutralization process.