[FIX] fix warning log level flag.
When using the flag "--log-level" with odev, it wasn't possible
to use the level of warning:

- "warn" is not a recognised log level in odev as defined in
odev/common/logging.py
- "warning" is not a recognised log level in Odoo, as per the
official Odoo documentation and the levels defined in odoo/odoo
/tools/config.py

Using the former in the --log-level flag terminates odev. Using
the latter doesn't work correctly, and prints all logs to the
standard output.

This commit addresses this issue by making warn a recognised
log level in odev.