[IMP] run: background-load second database from template while run
Before this commit:
Restoring a database template can be a long process depending on
the database size. In upgrade/maintenance, it can be needed that
we re-run the same database multiple times to test upgrade scripts.
As such, we generally interupt the running odoo-bin process then
launch again the commdand:
`odev run <db> -t <db template>`

After this commit:
The idea is to use the time while we are testing on the running
database to restore another database of the same template then
"swap" them when the first one is interupted in order to be
available to run as soon as possible.
This process goes on as long as odev itself is interupted with
Ctrl+C