[FIX] commands.database.create: fail if db not initialized
If the database was not initialized properly by odoo-bin odev should fail.
We then need to drop the database that was created and then exit.

We know the database was not initialized properly if OdoobinProcess.run raises an error OR
if its return value is None instead of being an instance of CompletedProcess(), which means the process caused an error but that error was catched.

fix for issue https://github.com/odoo-odev/odev/issues/66