[IMP] database.local: Handle invalid blcocks during filestore extraction
When run with Python 3.12 or lower, the zipfile module has a known
limitation that can lead to "invalid stored block lengths" errors when
extracting certain corrupted gzip files. This commit introduces a
warning message to inform users about this potential issue and suggests
upgrading to Python 3.13 or higher for improved handling of such cases.

To lower the impact of this issue, the code now attempts to skip the
extraction of corrupted files up to 10 times before aborting the entire
extraction process.