[IMP] run,create: support recursive addons paths
Find addons paths recursively when running a database, including
git submodules, to avoid having to list all submodules manually in the
command or configuration.

Also fixes a bug where addons paths were not correctly identified
when running a database where a save addons path was used but while not
in the current working directory.