[FIX] Dump local database with filestore now includes filestore files
The dump command for local databases was not including filestore files
in the dump archive, due to the archive being overriden with the SQL
dump of the database.

We are now making a single archive that includes both the SQL dump and
the filestore files in the format expected by Odoo.sh and Online.