[IMP] Install odev in its own virtual environment
Avoid errors on systems where installing python packages globally
is not allowed or discouraged (e.g Ubuntu 24.04).

Create a virtual environment in ~/.config/odev/venv and install
odev there. The odev.sh launcher script will use the virtualenv's
python interpreter if available, otherwise fallback to system's
python3 to keep compatibility with existing installs.

[IMP] Add migration instructions for switching to dedicated venv

[IMP] Update minium requirements and documentation

Update minimal python version to 3.10.12 and add a requirements section
in the README file.

Updaote the list of known plugins.

[IMP] Python and system packages installation improvements

Make it clearer which package managers were detected on the system
and raise an error if none is found when trying to install system
packages.

List packages about to be installed and let the user confirm before
proceeding.