[IMP] Retry dump restoration in degraded mode
When a Postgres dump restoration fails due to errors in the dump file
or local configuration, retry reimporting the same dump file ignoring
errors. This is slowers and could lead to data corruption
of the restored in some rare cases but will unblock users currently
unable to restore dump files. Occurring errors are logged at the end
of the process, if any.

This allows for removing the skip of specific SQL commands that was done
by parsing the dump file on the fly.