[FIX] Selection of specific commit SHA in worktree checkout/creation
We were prepending `origin/` in front of the commit SHA when selecting
a specific commit. While this helps with branches that are on the remote
but not fetched locally, this made it impossible to select a commit.

This commit fixes the issue by removing the `origin/` prefix when
the checked out revision is a commit SHA.