[ADD] Release channels (stable and beta) (#121)
* [ADD] Release channels (stable and beta)

Allow selection of a release channel for Odev and its plugins, directly
linked to an existing git branch.

Odev and official plugins will provide both 'main' and 'beta' branches.
Main (stable) will provide less frequent updates for the benefits of
less breaking changes. Beta will be used for more frequent updates but
will potentially break more often.

If repositories are set to track an unsupported branch, they will not
be switched automatically, this is to prevent loss of features
for people actively developping on Odev or a plugin.

* [FIX] Ignore release check on detached head

* [FIX] Handle dedent of None text